% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Graphs the population of a given country from 1950 to 2020}
\usage{
CountryPopulation(country)
}
\arguments{
\item{country}{A country in the WorldPopulation dataset}
}
\value{
The population graph of a country from 1950 to 2020
}
\description{
This function combines all of the year columns in the WorldPopulation dataset
for a given country and graphs the population for that country from 1950 to 2020.
}
\examples{
CountryPopulation("United States of America")
CountryPopulation("China")
}
